server/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
python-dotenv==1.0.0
